refactor: split FormDropdown into separate compact and search variants, move to Settings/common
